==> 1-vericekme.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import re

siteler = [
    "https://www.emlakjet.com/satilik-konut/istanbul-kadikoy/",
    "https://www.emlakjet.com/satilik-konut/istanbul-kadikoy/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-kadikoy/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-kartal/",
    "https://www.emlakjet.com/satilik-konut/istanbul-kartal/2/",
     "https://www.emlakjet.com/satilik-konut/istanbul-kartal/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-atasehir/",
     "https://www.emlakjet.com/satilik-konut/istanbul-atasehir/2/",
     "https://www.emlakjet.com/satilik-konut/istanbul-atasehir/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-besiktas/",
    "https://www.emlakjet.com/satilik-konut/istanbul-besiktas/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-besiktas/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-maltepe/",
    "https://www.emlakjet.com/satilik-konut/istanbul-maltepe/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-maltepe/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-pendik/",
    "https://www.emlakjet.com/satilik-konut/istanbul-pendik/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-pendik/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sancaktepe/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sancaktepe/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sancaktepe/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-uskudar/",
    "https://www.emlakjet.com/satilik-konut/istanbul-uskudar/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-uskudar/3/",
   
    "https://www.emlakjet.com/satilik-konut/istanbul-umraniye/",
    "https://www.emlakjet.com/satilik-konut/istanbul-umraniye/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-umraniye/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-tuzla/",
    "https://www.emlakjet.com/satilik-konut/istanbul-tuzla/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-tuzla/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-adalar/",
    "https://www.emlakjet.com/satilik-konut/istanbul-adalar/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-adalar/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-arnavutkoy/",
    "https://www.emlakjet.com/satilik-konut/istanbul-arnavutkoy/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-arnavutkoy/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-bahcelievler/",
    "https://www.emlakjet.com/satilik-konut/istanbul-bahcelievler/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-bahcelievler/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-bakirkoy/",
    "https://www.emlakjet.com/satilik-konut/istanbul-bakirkoy/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-bakirkoy/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-beykoz/",
    "https://www.emlakjet.com/satilik-konut/istanbul-beykoz/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-beykoz/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-avcilar/",
    "https://www.emlakjet.com/satilik-konut/istanbul-avcilar/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-avcilar/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-bagcilar/",
    "https://www.emlakjet.com/satilik-konut/istanbul-bagcilar/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-bagcilar/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-basaksehir/",
    "https://www.emlakjet.com/satilik-konut/istanbul-basaksehir/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-basaksehir/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-beylikduzu/",
    "https://www.emlakjet.com/satilik-konut/istanbul-beylikduzu/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-beylikduzu/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-buyukcekmece/",
    "https://www.emlakjet.com/satilik-konut/istanbul-buyukcekmece/2/",
     "https://www.emlakjet.com/satilik-konut/istanbul-buyukcekmece/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-cekmekoy/",
    "https://www.emlakjet.com/satilik-konut/istanbul-cekmekoy/2/",
     "https://www.emlakjet.com/satilik-konut/istanbul-cekmekoy/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-esenler/",
    "https://www.emlakjet.com/satilik-konut/istanbul-esenler/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-esenler/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-eyupsultan/",
    "https://www.emlakjet.com/satilik-konut/istanbul-eyupsultan/2/",
     "https://www.emlakjet.com/satilik-konut/istanbul-eyupsultan/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-gaziosmanpasa/",
    "https://www.emlakjet.com/satilik-konut/istanbul-gaziosmanpasa/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-gaziosmanpasa/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sile/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sile/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sile/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sultangazi/",
     "https://www.emlakjet.com/satilik-konut/istanbul-sultangazi/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sultangazi/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sultanbeyli/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sultanbeyli/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sultanbeyli/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-zeytinburnu/",
    "https://www.emlakjet.com/satilik-konut/istanbul-zeytinburnu/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-zeytinburnu/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-beyoglu/",
    "https://www.emlakjet.com/satilik-konut/istanbul-beyoglu/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-beyoglu/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sisli/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sisli/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sisli/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-esenyurt/",
    "https://www.emlakjet.com/satilik-konut/istanbul-esenyurt/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-esenyurt/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-fatih/",
    "https://www.emlakjet.com/satilik-konut/istanbul-fatih/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-fatih/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-gungoren/",
    "https://www.emlakjet.com/satilik-konut/istanbul-gungoren/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-gungoren/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-kagithane/",
    "https://www.emlakjet.com/satilik-konut/istanbul-kagithane/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-kagithane/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-kucukcekmece/",
    "https://www.emlakjet.com/satilik-konut/istanbul-kucukcekmece/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-kucukcekmece/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sariyer/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sariyer/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-sariyer/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-silivri/",
    "https://www.emlakjet.com/satilik-konut/istanbul-silivri/2/",
    "https://www.emlakjet.com/satilik-konut/istanbul-silivri/3/",
    "https://www.emlakjet.com/satilik-konut/istanbul-catalca/",
    "https://www.emlakjet.com/satilik-konut/istanbul-catalca/2/",
     "https://www.emlakjet.com/satilik-konut/istanbul-catalca/3/"



    

]
emlakjetİlan = []

for site_url in siteler:
    # Sayfa içeriğini alma
    site = requests.get(site_url)
    soup = BeautifulSoup(site.content, "lxml")
    st1 = soup.find("div", {"class": "ej64 ej100 ej156 _2g-Ygo"})
    st2 = st1.find_all("div", {"class": "_3qUI9q"})

    # Her bir ilan için işlem
    for icerik in st2:
        link_sonu = icerik.a.get("href")
        link_basi = "https://www.emlakjet.com"
        link = link_basi + link_sonu
        detay = requests.get(link)
        detay_soup = BeautifulSoup(detay.content, "lxml")
        ilan_ad = detay_soup.find('div', class_="_3VQ1JB")

        # Eğer ilan adı bulunduysa, içerisindeki metni al
        if ilan_ad:
            metin = ilan_ad.find('p').get_text()
            # Metni "-" karakterine göre böler ve istediğimiz parçayı alırız
            bolunmus_metin = metin.split(" - ")
            if len(bolunmus_metin) >= 2:
                istenilen_kisim = bolunmus_metin[1]  # "Kadıköy" kısmı
                
            else:
                continue  # Gerekli bilgi yoksa, bu ilanı atlar

            # İlan fiyatını alırken düzenli ifade kullanma
            ilan_fiyat_etiketi = detay_soup.find("p", {"class": "_1NQ0d5"})
            if ilan_fiyat_etiketi:
                ilan_fiyat = ilan_fiyat_etiketi.text.strip()
                sadece_sayi = re.findall(r'\d+', ilan_fiyat)
                sayi = ''.join(sadece_sayi)
                ilan_fiyat = sayi
            else:
                continue  # İlan fiyatı yoksa, bu ilanı atlar

            tum_etiketler = detay_soup.find_all("div", {"class": "_35T4WV"})
            ilan_detay = {}

            for etiket in tum_etiketler:
                anahtar = etiket.find("div", {"class": "_1bVOdb"}).text.strip()
                deger = etiket.find_all("div", {"class": "_1bVOdb"})[1].text.strip()
                ilan_detay[anahtar] = deger

            emlakjetİlan.append({"ilan_ilce": istenilen_kisim, "ilan_fiyat": ilan_fiyat, **ilan_detay})
            
# Ana ilan bilgilerini içeren DataFrame oluşturma
df_ilanlar = pd.DataFrame(emlakjetİlan)

# CSV dosyasına yaz
df_ilanlar.to_csv("emlakjetİlan.csv", index=False, encoding="utf-8")
