==> 3-onisleme.ipynb <==
import pandas as pd

df = pd.read_csv("emlakjet.csv")

df

from sklearn import preprocessing 
le = preprocessing.LabelEncoder()

df_2 = df.copy()
df_3 = df.copy()

df_3["ilan_ilce"] = le.fit_transform(df_2.ilan_ilce)

df_3["oda_sayısı"] = le.fit_transform(df_2.oda_sayısı)

df_3["bulunduğu_kat"] = le.fit_transform(df_2.bulunduğu_kat)

df_3["ısıtma_tipi"] = le.fit_transform(df_2.ısıtma_tipi)

df_3["krediye_uygunluk"] = le.fit_transform(df_2.krediye_uygunluk)

df_3["banyo_sayısı"] = le.fit_transform(df_2.banyo_sayısı)

df_3["tipi"] = le.fit_transform(df_2.tipi)

df_3["binanın_yaşı"] = le.fit_transform(df_2.binanın_yaşı)

df_3["binanın_kat_sayısı"] = le.fit_transform(df_2.binanın_kat_sayısı)

df_3["binanın_yaşı"] = le.fit_transform(df_2.binanın_yaşı)

df_3["binanın_kat_sayısı"] = le.fit_transform(df_2.binanın_kat_sayısı)

df_3["site_içerisinde"] = le.fit_transform(df_2.site_içerisinde)

df_3.to_csv("emlakjetisleme.csv")

df = pd.read_csv("emlakjetisleme.csv")

df['net_metrekare'] = df['net_metrekare'].astype(int)
df['brüt_metrekare'] = df['brüt_metrekare'].astype(int)


df.info()
