==> 5-)Arayuz.ipynb <==
from tkinter import *
from tkinter import messagebox
from tkinter.ttk import Combobox
import pandas as pd
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.ensemble import GradientBoostingRegressor

# Veri setini oku
df = pd.read_csv("emlakjetisleme.csv")

# Bağımlı ve bağımsız değişkenlerin ayrılması
X = df.drop(["ilan_fiyat", ], axis=1)
y = df["ilan_fiyat"]
# Eğitim verilerini ve test verilerini ayırma
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=144)


# Parametre aralıklarını belirleme
param_grid = {
    'n_estimators': [100, 200, 300],
    'learning_rate': [0.05, 0.1, 0.2],
    'max_depth': [3, 4, 5]
}
# Gradient Boosting Regressor modeli
model_gb = GradientBoostingRegressor(random_state=144)

# Parametre optimizasyonu için GridSearchCV kullanma
grid_search = GridSearchCV(estimator=model_gb, param_grid=param_grid, cv=3, n_jobs=-1, scoring='r2')
grid_search.fit(X_train, y_train)
# En iyi parametreler
best_params = grid_search.best_params_
print("En iyi parametreler:", best_params)

# En iyi modeli seçme
best_model_gb = grid_search.best_estimator_
# Modelin performansını test verileri ile kontrol etme
best_model_score = best_model_gb.score(X_test, y_test)
print("Model Score:", best_model_score)

# Pencere ayarları
pencere = Tk()
pencere.title("Konut Fiyat Tahmini")
pencere.configure(background='#E0B0FF')
pencere.geometry("800x600")

def mesaj():
    messagebox.showinfo(title="Başarılı", message="Seçim başarılı")

def olumsuz():
    messagebox.showwarning(title="Dikkat", message="Seçim Yapmadınız")

# Değişkenler
# Değişkenleri tanımlama
ilce, tür, net, oda, yaş, site, alan, ısıtma, banyo, daire, kredi = None, None, None, None, None, None, None, None, None, None, None


def ilce_düzenle(event):
    global ilce 
    ilce_deger = ilce_kutu.get()
    ilceler = {"Kadıköy": 19, "Kartal": 20, "Ataşehir": 2, "Beşiktaş": 11, "Maltepe": 23, "Pendik": 24,
               "Sancaktepe": 25, "Üsküdar": 35, "Ümraniye": 34, "Tuzla": 30, "Adalar": 0, "Arnavutköy": 1,
               "Bahçelievler": 4, "Bakırköy": 5, "Beykoz": 8, "Avcılar": 3, "Bağcılar": 6, "Başakşehir": 7,
               "Beylikdüzü": 9, "Büyükçekmece": 12, "Çekmeköy": 33, "Esenler": 13, "Eyüpsultan": 15,
               "Gaziosmanpaşa": 17, "Şile": 36, "Sultangazi": 29, "Sultanbeyli": 28, "Zeytinburnu": 31,
               "Beyoğlu": 10, "Şişli": 37, "Esenyurt": 14, "Fatih": 16, "Güngören": 18, "Kağıthane": 21,
               "Küçükçekmece": 22, "Sarıyer": 26, "Silivri": 27, "Çatalca": 32}
    ilce = ilceler.get(ilce_deger, None)
    if ilce is not None:
        mesaj()
    else:
        olumsuz()

def tür_düzenle(event):
    global tür
    tür_deger = tür_kutu.get()
    türler = {"Daire": 1, "Villa": 5, "Müstakil Ev": 2, "Bina": 0, "Residence": 4, "Yazlık": 6, "Prefabrik": 3}
    tür = türler.get(tür_deger, None)
    if tür is not None:
        mesaj()
    else:
        olumsuz()

def net_düzenle(event):
    global net
    net_metrekare = int(net_entry.get())
    if net_metrekare > 0:
        net = net_metrekare
        mesaj()
    else:
        olumsuz()

def oda_düzenle(event):
    global oda
    oda_deger = oda_kutu.get()
    odalar = {"1+1": 0, "2+1": 1, "2+2": 2, "3+1": 3, "3+2": 4, "4+1": 5, "4+2": 6, "5+1": 7, "5+2": 8, "6+1": 9, 
              "6+2": 10, "7+1": 11, "9+": 12}
    oda = odalar.get(oda_deger, None)
    if oda is not None:
        mesaj()
    else:
        olumsuz()

def yaş_düzenle(event):
    global yaş
    yaş_deger = yaslar_kutu.get()
    yaşlar = {"0": 0, "1": 1, "2": 4, "3": 6, "4": 7, "5-10": 8, "11-15": 2, "16-20": 3, "21 ve üzeri": 5}
    yaş = yaşlar.get(yaş_deger, None)
    if yaş is not None:
        mesaj()
    else:
        olumsuz()

def site_düzenle(event):
    global site
    site_deger = site_kutu.get()
    if site_deger == "Evet":
        site = 0
        mesaj()
    elif site_deger == "Hayır":
        site = 1
        mesaj()
    else:
        olumsuz()
def binakat_düzenle(event):
    global binakat
    binakat_deger = binakat_kutu.get()
    if binakat_deger == "1":
        binakat = 0
        mesaj()
    elif binakat_deger == "2":
        binakat = 1
        mesaj()
    elif binakat_deger == "3":
        binakat = 2
        mesaj()
    elif binakat_deger == "4":
        binakat = 3
        mesaj()
    elif binakat_deger == "5":
        binakat = 4
        mesaj()
    elif binakat_deger == "6":
        binakat = 5
        mesaj()
    elif binakat_deger == "7":
        binakat = 6
        mesaj()
    elif binakat_deger == "8":
        binakat = 7
        mesaj()
    elif binakat_deger == "9":
        binakat = 8
        mesaj()
    elif binakat_deger == "10":
        binakat = 9
        mesaj()
    elif binakat_deger == "15":
        binakat = 10
        mesaj()
    elif binakat_deger == "20":
        binakat = 11
        mesaj()
    elif binakat_deger == "30":
        binakat = 12
        mesaj()
    elif binakat_deger == "40":
        binakat = 13
        mesaj()
    else:
        olumsuz()


def kredi_düzenle(event):
    global kredi
    kredi_deger = kredi_kutu.get()
    if kredi_deger == "Krediye Uygun":
        kredi = 0
        mesaj()
    elif kredi_deger == "Krediye Uygun Değil":
        kredi = 1
        mesaj()
    else:
        olumsuz()

def alan_düzenle(event):
    global alan
    alan_ent = int(alan_entry.get())
    if alan_ent > 0:
        alan = alan_ent
        mesaj()
    else:
        olumsuz()

def ısıtma_düzenle(event):
    global ısıtma
    ısıtma_deger = ısıtma_kutu.get()
    ısıtmalar = {"Doğalgaz": 0, "Isıtma Yok": 1, "Klimalı": 2, "Merkezi": 3, "Şömine": 6, "Sobalı": 4, "Yerden Isıtma": 5}
    ısıtma = ısıtmalar.get(ısıtma_deger, None)
    if ısıtma is not None:
        mesaj()
    else:
        olumsuz()

def banyo_düzenle(event):
    global banyo
    banyo_deger = banyo_kutu.get()
    banyolar = {"Yok": 6, "1": 0, "2": 1, "3": 2, "4": 3, "5": 4, "6+": 5}
    banyo = banyolar.get(banyo_deger, None)
    if banyo is not None:
        mesaj()
    else:
        olumsuz()

def daire_düzenle(event):
    global daire
    daire_deger = daire_kutu.get()
    daireler = {"-1": 0, "-2": 1, "Giriş": 18, "1": 2, "2": 6, "3": 9, "4": 11, "5": 12, "6": 13, "7": 14, 
                "8": 15, "9": 16, "10": 3, "11-15": 4, "16-20": 5, "21-25": 7, "26-30": 8, "31-38": 10,
                "Bahçe Katı": 17, "Müstakil": 19, "Çatı Dublex": 20, "Çatı Katı": 21, "Zemin": 22, "Teras Katı": 23}
    daire = daireler.get(daire_deger, None)
    if daire is not None:
        mesaj()
    else:
        olumsuz()

def predict_price():
    if None in (ilce, net, oda, daire, ısıtma, kredi, banyo, tür, alan, yaş, daire,binakat, site):
        olumsuz()
        return

    user_input = pd.DataFrame([{
        'ilan_ilce': ilce,
        'net_metrekare': net,
        'oda_sayısı': oda,
        'bulunduğu_kat': daire,
        'ısıtma_tipi': ısıtma,
        'krediye_uygunluk': kredi,
        'banyo_sayısı': banyo,
        'tipi': tür,
        'brüt_metrekare': alan,
        'binanın_yaşı': yaş,
        'binanın_kat_sayısı':binakat,
        'site_içerisinde': site,
    }])

    prediction = best_model_gb.predict(user_input)
    messagebox.showinfo(title="Tahmin", message=f"Tahmin Edilen Fiyat: {prediction[0]:,.2f} TL")

# Kullanıcı arayüzü
ilce_label = Label(pencere, text="İlçe:", bg='#E0B0FF')
ilce_label.grid(row=0, column=0, padx=10, pady=5)
ilce_kutu = Combobox(pencere)
ilce_kutu['values'] = list({"Kadıköy", "Kartal", "Ataşehir", "Beşiktaş", "Maltepe", "Pendik",
                            "Sancaktepe", "Üsküdar", "Ümraniye", "Tuzla", "Adalar", "Arnavutköy",
                            "Bahçelievler", "Bakırköy", "Beykoz", "Avcılar", "Bağcılar", "Başakşehir",
                            "Beylikdüzü", "Büyükçekmece", "Çekmeköy", "Esenler", "Eyüpsultan",
                            "Gaziosmanpaşa", "Şile", "Sultangazi", "Sultanbeyli", "Zeytinburnu",
                            "Beyoğlu", "Şişli", "Esenyurt", "Fatih", "Güngören", "Kağıthane",
                            "Küçükçekmece", "Sarıyer", "Silivri", "Çatalca"})
ilce_kutu.grid(row=0, column=1, padx=10, pady=5)
ilce_kutu.bind("<<ComboboxSelected>>", ilce_düzenle)

tür_label = Label(pencere, text="Konut Türü:", bg='#E0B0FF')
tür_label.grid(row=1, column=0, padx=10, pady=5)
tür_kutu = Combobox(pencere)
tür_kutu['values'] = list({"Daire", "Villa", "Müstakil Ev", "Bina", "Residence", "Yazlık", "Prefabrik"})
tür_kutu.grid(row=1, column=1, padx=10, pady=5)
tür_kutu.bind("<<ComboboxSelected>>", tür_düzenle)

net_label = Label(pencere, text="Net Metrekare:", bg='#E0B0FF')
net_label.grid(row=2, column=0, padx=10, pady=5)
net_entry = Entry(pencere)
net_entry.grid(row=2, column=1, padx=10, pady=5)
net_entry.bind("<Return>", net_düzenle)

oda_label = Label(pencere, text="Oda Sayısı:", bg='#E0B0FF')
oda_label.grid(row=3, column=0, padx=10, pady=5)
oda_kutu = Combobox(pencere)
oda_kutu['values'] = list({"1+1", "2+1", "2+2", "3+1", "3+2", "4+1", "4+2", "5+1", "5+2", "6+1", 
                           "6+2", "7+1", "9+"})
oda_kutu.grid(row=3, column=1, padx=10, pady=5)
oda_kutu.bind("<<ComboboxSelected>>", oda_düzenle)

yaş_label = Label(pencere, text="Bina Yaşı:", bg='#E0B0FF')
yaş_label.grid(row=4, column=0, padx=10, pady=5)
yaslar_kutu = Combobox(pencere)
yaslar_kutu['values'] = list({"0", "1", "2", "3", "4", "5-10", "11-15", "16-20", "21 ve üzeri"})
yaslar_kutu.grid(row=4, column=1, padx=10, pady=5)
yaslar_kutu.bind("<<ComboboxSelected>>", yaş_düzenle)

site_label = Label(pencere, text="Site İçerisinde:", bg='#E0B0FF')
site_label.grid(row=5, column=0, padx=10, pady=5)
site_kutu = Combobox(pencere)
site_kutu['values'] = list({"Evet", "Hayır"})
site_kutu.grid(row=5, column=1, padx=10, pady=5)
site_kutu.bind("<<ComboboxSelected>>", site_düzenle)

alan_label = Label(pencere, text="Brüt Metrekare:", bg='#E0B0FF')
alan_label.grid(row=6, column=0, padx=10, pady=5)
alan_entry = Entry(pencere)
alan_entry.grid(row=6, column=1, padx=10, pady=5)
alan_entry.bind("<Return>", alan_düzenle)

ısıtma_label = Label(pencere, text="Isıtma Tipi:", bg='#E0B0FF')
ısıtma_label.grid(row=7, column=0, padx=10, pady=5)
ısıtma_kutu = Combobox(pencere)
ısıtma_kutu['values'] = list({"Doğalgaz", "Isıtma Yok", "Klimalı", "Merkezi", "Şömine", "Sobalı", "Yerden Isıtma"})
ısıtma_kutu.grid(row=7, column=1, padx=10, pady=5)
ısıtma_kutu.bind("<<ComboboxSelected>>", ısıtma_düzenle)

banyo_label = Label(pencere, text="Banyo Sayısı:", bg='#E0B0FF')
banyo_label.grid(row=8, column=0, padx=10, pady=5)
banyo_kutu = Combobox(pencere)
banyo_kutu['values'] = list({"Yok", "1", "2", "3", "4", "5", "6+"})
banyo_kutu.grid(row=8, column=1, padx=10, pady=5)
banyo_kutu.bind("<<ComboboxSelected>>", banyo_düzenle)

daire_label = Label(pencere, text="Dairenin Katı:", bg='#E0B0FF')
daire_label.grid(row=9, column=0, padx=10, pady=5)
daire_kutu = Combobox(pencere)
daire_kutu['values'] = list({"-1", "-2", "Giriş", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10", 
                             "11-15", "16-20", "21-25", "26-30", "31-38", "Bahçe Katı", "Müstakil", 
                             "Çatı Dublex", "Çatı Katı", "Zemin", "Teras Katı"})
daire_kutu.grid(row=9, column=1, padx=10, pady=5)
daire_kutu.bind("<<ComboboxSelected>>", daire_düzenle)

binakat_label = Label(pencere, text="Binanın Kat Sayısı:", bg='#E0B0FF')
binakat_label.grid(row=10, column=0, padx=10, pady=5)
binakat_kutu = Combobox(pencere)
binakat_kutu['values'] = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 15, 20, 30, 40]

binakat_kutu.grid(row=10, column=1, padx=10, pady=5)
binakat_kutu.bind("<<ComboboxSelected>>", binakat_düzenle)


kredi_label = Label(pencere, text="Kredi Durumu:", bg='#E0B0FF')
kredi_label.grid(row=11, column=0, padx=10, pady=5)
kredi_kutu = Combobox(pencere)
kredi_kutu['values'] = list({"Krediye Uygun", "Krediye Uygun Değil"})
kredi_kutu.grid(row=11, column=1, padx=10, pady=5)
kredi_kutu.bind("<<ComboboxSelected>>", kredi_düzenle)

predict_button = Button(pencere, text="Tahmin Et", command=predict_price, bg='#A020F0', fg='white')
predict_button.grid(row=12, column=0, columnspan=2, pady=20)

# Mainloop
pencere.mainloop()
